requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
install==1.3.5
pillow==10.3.0
psycopg==3.1.11
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
